split sundry api
